refactor(CampaignsHistory): change the folder and names
