+++ src/App.js
@@ -15,7 +15,7 @@ import { availableLanguageOrNull } from './i18n/utils';
 import Shopify from './components/Integrations/Shopify/Shopify';
 import SignupConfirmation from './components/Signup/SignupConfirmation';
 import SubscriberHistory from './components/Reports/SubscriberHistory/SubscriberHistory';
-import SubscriberHistorySentCampaigns from './components/Reports/SubscriberHistory/SubscriberHistoryCurrentSearch/SubscriberHistorySentCampaigns/SubscriberHistorySentCampaigns';
+import CampaignsHistory from './components/Reports/CampaignsHistory/CampaignsHistory';
 
 /**
  * @param { Object } props - props
@@ -92,11 +92,7 @@ const App = ({ locale, location, dependencies: { appSessionRef, sessionManager }
           <PrivateRoute path="/reports/" exact requireSiteTracking component={Reports} />
           <PrivateRoute path="/integrations/shopify" exact component={Shopify} />
           <PrivateRoute path="/reports/subscriber-history" exact component={SubscriberHistory} />
-          <PrivateRoute
-            path="/reports/campaigns-history"
-            exact
-            component={SubscriberHistorySentCampaigns}
-          />
+          <PrivateRoute path="/reports/campaigns-history" exact component={CampaignsHistory} />
           <PublicRouteWithLegacyFallback exact path="/login" />
           <PublicRouteWithLegacyFallback exact path="/signup" />
           <PublicRouteWithLegacyFallback exact path="/login/reset-password" />
